requirement.txt
altgraph==0.14
bleach==2.1.1
colorama==0.3.9
cycler==0.10.0
decorator==4.1.2
entrypoints==0.2.3
future==0.16.0
html5lib==1.0b10
ipykernel==4.6.1
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.0.2
jedi==0.11.0
Jinja2==2.9.6
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
macholib==1.8
MarkupSafe==1.0
matplotlib==2.1.0
mistune==0.7.4
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.2.0
numpy==1.13.3
pandocfilters==1.4.2
parso==0.1.0
pefile==2017.9.3
pickleshare==0.7.4
prompt-toolkit==1.0.15
Pygments==2.2.0
PyInstaller==3.3
pyparsing==2.2.0
PyQt5==5.9
pyqt5-tools==5.9.0.1.2
python-dateutil==2.6.1
pytz==2017.2
PyVISA==1.8
pywin32==221
pyzmq==16.0.2
qtconsole==4.3.1
simplegeneric==0.8.1
sip==4.19.3
six==1.11.0
testpath==0.3.1
tornado==4.5.2
traitlets==4.3.2
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.0.4
wxPython==4.0.0b2
